[c2842] Implement script for Scourgelord Tyrannus - Pit of Saron
This will conclude the Pit of Saron implementation. There are however a few small things which still need to be implemented or fixed, but the main instance events are fully playable.

Thank you for choosing SD2 and please don't forget to report any issues that you find!

(based on commit [2842] - a2c53e4)

Signed-off-by: Schmoozerd <schmoozerd@scriptdev2.com>